Move the imaging memory policy off the settings class.
ImagingMemorySettings carried the derivation as well as the configured values:
six public methods and ten private constants, all of it permanent API surface
for the major once it ships. None of it means anything on its own, either -
what a zero derives to, and whether a bound applies at all, is a judgement
about ImageSharp's own defaults, which Umbraco.Core knows nothing about.

Move all of it to ImageProcessingMemory, which both imaging packages already
share as a linked source file, and have it read the settings passed in. The
settings class keeps the four properties, their defaults, and the ceiling the
validator enforces - that one constrains the setting rather than deriving from
it, and the validator lives in Core. Nothing about the derivation is public any
longer, so none of it is locked for the major. The tests move with it, into the
fixture for the class that now owns the behaviour.
